oneplus2: remove scaling_max_freq and save restore for 8994
A57 frequency needs to be limited during boot when thermal hotplug is
disabled. We achieve this by using msm_performance module.

Remove save and restore of scaling_max_freq as it's redundant and
error-prone due to the fact that scaling_max_freq node won't be
available if a core is hotplugged out.

Change-Id: I69cecbcc6c219d4c428add993a32748a460953ef
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>